#18 シリアライザを分割（post, post_image, post_prefecture）
