+++ fosterParentDog/serializers/post_serializer.go
@@ -64,55 +64,3 @@ func (ps *PostSerializer) Response() []PostResponse {
 	}
 	return retSlice
 }
-
-//*******************************************************************
-// PostImage
-//*******************************************************************
-type PostImageSerializer struct {
-	C               *gin.Context
-	EntityPostImage []entity.PostImage
-}
-
-type PostImageResponse struct {
-	PostID      uint   `json:"post_id"`
-	PostImageID uint   `json:"post_image_id"`
-	ImagePath   string `json:"image_path"`
-}
-
-func (ps *PostImageSerializer) ResponsePostImage() []PostImageResponse {
-	myModel := ps.C.MustGet("my_post_image_model").([]entity.PostImage)
-	var postRet []PostImageResponse
-	for i := 0; i < len(myModel); i++ {
-		postRet = append(postRet, PostImageResponse{
-			PostID:      ps.EntityPostImage[i].PostId,
-			PostImageID: ps.EntityPostImage[i].ID,
-			ImagePath:   ps.EntityPostImage[i].ImagePath,
-		})
-	}
-	return postRet
-}
-
-//*******************************************************************
-// PostTransferablePrefecture
-//*******************************************************************
-type PostPrefectureSerializer struct {
-	C                    *gin.Context
-	EntityPostPrefecture []entity.PostPrefecture
-}
-
-type PostPrefectureResponse struct {
-	PostId           uint `json:"post_id"`
-	PostPrefectureId int  `json:"post_prefecture_id"`
-}
-
-func (ps *PostPrefectureSerializer) ResponsePostPrefecture() []PostPrefectureResponse {
-	myModel := ps.C.MustGet("my_post_prefecture_model").([]entity.PostPrefecture)
-	var postRet []PostPrefectureResponse
-	for i := 0; i < len(myModel); i++ {
-		postRet = append(postRet, PostPrefectureResponse{
-			PostId:           ps.EntityPostPrefecture[i].PostId,
-			PostPrefectureId: ps.EntityPostPrefecture[i].PostPrefectureId,
-		})
-	}
-	return postRet
-}
